// frontend/src/lib/api/lasius/webSocketIssueImporterSyncStatsChanged.ts
/**
 * Lasius - Open source time tracker for teams
 * Copyright (c) Tegonal Genossenschaft (https://tegonal.com)
 *
 * This file is part of Lasius.
 *
 * Lasius is free software: you can redistribute it and/or modify it under the terms of the
 * GNU Affero General Public License as published by the Free Software Foundation, either
 * version 3 of the License, or (at your option) any later version.
 *
 * Lasius is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
 * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License along with Lasius.
 * If not, see <https://www.gnu.org/licenses/>.
 *
 */

import type { ModelsConfigSyncStatus } from './modelsConfigSyncStatus'
import type { ModelsImporterType } from './modelsImporterType'
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lasius API
 * Track your time
 * OpenAPI spec version: 2.0.1+10-a10445d4+20251012-1746
 */
import type { WebSocketIssueImporterSyncStatsChangedType } from './webSocketIssueImporterSyncStatsChangedType'

/**
 * Issue importer sync status changed (admin-only, negative outcomes only)
 */
export interface WebSocketIssueImporterSyncStatsChanged {
  type: WebSocketIssueImporterSyncStatsChangedType
  configId: string
  organisationId: string
  importerType: ModelsImporterType
  configName: string
  syncStatus: ModelsConfigSyncStatus
}